Adding flows that assess the correctness of the major types of automations
For each of these cases, we (re)create an automation that is looking for very
specific conditions that are unique to the test (using a unique resource ID).
We then emit the events necessary to make the automation fire, then wait for it
to be triggered (or timeout and fail if it doesn't happen in time).

Note: I wasn't sure how to make an effective integration test for metric
triggers at this time, given the longer time horizons.  This is something we
can work on over time.