Simplify the notification messages from integration tests
We were using a fairly chatty and dense format for these notifications, and we
were using the `{{ flow_run }}` variable, which may occasionally lead to some
misleading information in the message (if it takes longer than usual to trigger
the automation).

This changes the notifications to be more human readable and to use only the
information embedded within the triggering event itself to make sure it's the
most accurate.

Helps improve the situation of #7676