Use concurrency limits to only allow a single automation integration test at once (#8)
When we have a situation where the workers in staging are offline, or other
interruptions disrupt the scheduling of these automations, we can end up with
multiple runs overlapping.  When that happens, the automations may misfire
because a subsequent run has recreated the automation.

Using a concurrency limit of `1` should help these from stomping on themselves.