Increases the event visibility timeouts to 180s (#21)
We are seeing elevated latency on the `events-websockets` in staging since we've
upgraded to Prefect 3.0 (with CSTRO and with distributed futures).  Until we
have a fix for reducing the latency of the outbound websockets, we will increase
the timeouts of the final event checks of these automations tests to accomodate.